refactor(gdb): lv_global supports lazy loading and refactors some structures and formats.
