+++ scripts/gdb/lvglgdb/gdb_cmds/__init__.py
@@ -1,0 +1,11 @@
+from .core import DumpObj
+from .display import DumpDisplayBuf
+from .draw import InfoDrawUnit
+from .misc import InfoStyle
+
+__all__ = [
+    "DumpObj",
+    "DumpDisplayBuf",
+    "InfoDrawUnit",
+    "InfoStyle",
+]
