+++ scripts/gdb/lvglgdb/gdb_cmds/core/__init__.py
@@ -1,0 +1,5 @@
+from .lv_obj import DumpObj
+
+__all__ = [
+    "DumpObj",
+]
