+++ scripts/gdb/lvglgdb/gdb_cmds/display/__init__.py
@@ -1,0 +1,5 @@
+from .lv_display import DumpDisplayBuf
+
+__all__ = [
+    "DumpDisplayBuf",
+]
