+++ scripts/gdb/lvglgdb/gdb_cmds/draw/__init__.py
@@ -1,0 +1,5 @@
+from .lv_draw import InfoDrawUnit
+
+__all__ = [
+    "InfoDrawUnit",
+]
