+++ scripts/gdb/lvglgdb/gdb_cmds/misc/__init__.py
@@ -1,0 +1,5 @@
+from .lv_style import InfoStyle
+
+__all__ = [
+    "InfoStyle",
+]
