+++ scripts/gdb/lvglgdb/lvgl/__init__.py
@@ -1,0 +1,14 @@
+from .core import LVObject, curr_inst, dump_obj_info
+from .display import LVDisplay
+from .draw import LVDrawBuf
+from .misc import LVList, dump_style_info
+
+__all__ = [
+    "LVObject",
+    "LVDisplay",
+    "LVDrawBuf",
+    "curr_inst",
+    "LVList",
+    "dump_style_info",
+    "dump_obj_info",
+]
