+++ scripts/gdb/lvglgdb/lvgl/core/__init__.py
@@ -1,0 +1,8 @@
+from .lv_obj import LVObject, dump_obj_info
+from .lv_global import curr_inst
+
+__all__ = [
+    "LVObject",
+    "curr_inst",
+    "dump_obj_info",
+]
