+++ scripts/gdb/lvglgdb/lvgl/core/lv_obj.py
@@ -1,8 +1,4 @@
-import argparse
-import gdb
-
-from ..value import Value
-from .. import lv_global
+from lvglgdb.value import Value
 
 
 class LVObject(Value):
@@ -68,64 +64,6 @@ class LVObject(Value):
 
     def get_child(self, index: int):
         return self.spec_attr.children[index] if self.spec_attr else None
-
-
-class DumpObj(gdb.Command):
-    """dump obj tree from specified obj"""
-
-    def __init__(self):
-        super(DumpObj, self).__init__(
-            "dump obj", gdb.COMMAND_USER, gdb.COMPLETE_EXPRESSION
-        )
-
-    def dump_obj(self, obj: LVObject, depth=0, limit=None):
-        if not obj:
-            return
-
-        # dump self
-        print("  " * depth, end="")
-        dump_obj_info(obj)
-
-        if limit is not None and depth >= limit:
-            return
-
-        # dump children
-        for child in obj.children:
-            self.dump_obj(child, depth + 1, limit=limit)
-
-    def invoke(self, args, from_tty):
-        parser = argparse.ArgumentParser(description="Dump lvgl obj tree.")
-        parser.add_argument(
-            "-L",
-            "--level",
-            type=int,
-            default=None,
-            help="Limit the depth of the tree.",
-        )
-        parser.add_argument(
-            "root",
-            type=str,
-            nargs="?",
-            default=None,
-            help="Optional root obj to dump.",
-        )
-        try:
-            args = parser.parse_args(gdb.string_to_argv(args))
-        except SystemExit:
-            return
-
-        if args.root:
-            root = gdb.parse_and_eval(args.root)
-            root = LVObject(root)
-            self.dump_obj(root, limit=args.level)
-        else:
-            # dump all displays
-            depth = 0
-            for disp in lv_global.g_lvgl_instance.displays():
-                print(f"Display {hex(disp)}")
-                for screen in disp.screens:
-                    print(f'{"  " * (depth + 1)}Screen@{hex(screen)}')
-                    self.dump_obj(screen, depth=depth + 1, limit=args.level)
 
 
 def dump_obj_info(obj: LVObject):
