+++ scripts/gdb/lvglgdb/lvgl/display/__init__.py
@@ -1,0 +1,5 @@
+from .lv_display import LVDisplay
+
+__all__ = [
+    "LVDisplay",
+]
