+++ scripts/gdb/lvglgdb/lvgl/draw/__init__.py
@@ -1,0 +1,5 @@
+from .lv_draw_buf import LVDrawBuf
+
+__all__ = [
+    "LVDrawBuf",
+]
